+++ tests/integration/core_test.py
@@ -4,10 +4,11 @@ import sys
 from dotenv import load_dotenv
 
 # Add the `../src` directory to sys.path
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../src")))
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../src")))
 
 # Now you can import the module
 from xurrent.core import XurrentApiHelper
+from xurrent.teams import Team
 from xurrent.people import Person
 
 # FILE: src/xurrent/test_core.py
@@ -40,5 +41,10 @@ def test_api_helper_setup(x_api_helper):
     assert isinstance(x_api_helper.api_user, Person)
     assert x_api_helper.api_user.id is not None
     assert isinstance(x_api_helper.api_user.id, int)
+
+    #check if the api_user_teams is a list
+    assert isinstance(x_api_helper.api_user_teams, list)
+    for team in x_api_helper.api_user_teams:
+        assert isinstance(team, Team)